requirements.txt
astunparse==1.6.3
contourpy==1.3.2
cycler==0.12.1
fonttools==4.61.1
jittor==1.3.8.5
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==1.26.4
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079158/work
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scipy==1.15.3
six==1.17.0
tqdm==4.67.3
